third_party/WebKit/LayoutTests/fast/scroll-behavior/overflow-scroll-animates-expected.txt

Tests that overflow scrolls are animated, unless the wheel event has precise scrolling deltas.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.scrollTop is not 80
PASS element.scrollTop is 80
PASS successfullyParsed is true

TEST COMPLETE
